backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.28.1

# Scheduler
apscheduler==3.10.4

# ML Inference
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3

# File uploads
aiofiles==24.1.0
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.22.1
